requirements.txt
Django>=5.0
celery[redis]>=5.3
django-celery-beat>=2.5
django-redis>=5.4
playwright==1.54.0
psycopg2-binary>=2.9
django-unfold
python-dotenv~=1.1.1
asyncio~=4.0.0
unfold~=1.2.3
djangorestframework~=3.16.1
requests~=2.32.4
python-telegram-bot~=21.0
pytz>=2023.3
